refactor(cef): split client/mod.rs into focused submodules
