+++ .changesets/1782973032-refactor-cef-split-client-mod-rs-into-submodules-7308.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): split client/mod.rs into submodules
